pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Aleksandr Vednikov <aleasterxs@mail.ru>"]
readme = "README.md"
packages = [
    {include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.12"
argparse = "^1.4.0"
pyyaml = "^6.0.1"
pathlib = "^1.0.1"


[tool.poetry.scripts]
gendiff = "gendiff.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
